handle deepgram results with provider
