+++ apps/course-builder-web/src/inngest/video-processing/order-transcript.ts
@@ -1,16 +1,8 @@
-import { env } from '@/env.mjs'
 import { inngest } from '@/inngest/inngest.server'
 import { VIDEO_RESOURCE_CREATED_EVENT } from '@/inngest/video-processing/events/video-resource'
 import { getVideoResource } from '@/lib/video-resource-query'
+import { transcriptProvider } from '@/providers/deepgram'
 import { NonRetriableError } from 'inngest'
-
-import Deepgram from '@coursebuilder/core/providers/deepgram'
-
-const callbackBase = env.NODE_ENV === 'production' ? env.UPLOADTHING_URL : env.NEXT_PUBLIC_URL
-const deepgramService = Deepgram({
-  apiKey: env.DEEPGRAM_API_KEY,
-  callbackUrl: `${callbackBase}/api/deepgram/webhook`,
-})
 
 export const orderTranscript = inngest.createFunction(
   { id: `order-transcript`, name: 'Order Transcript from Deepgram' },
@@ -25,7 +17,7 @@ export const orderTranscript = inngest.createFunction(
     }
 
     const deepgram = await step.run('Order Transcript [Deepgram]', async () => {
-      return await deepgramService.initiateTranscription({
+      return await transcriptProvider.initiateTranscription({
         mediaUrl: event.data.originalMediaUrl,
         resourceId: event.data.videoResourceId,
       })
