+++ packages/core/package.json
@@ -48,6 +48,7 @@
   "dependencies": {},
   "devDependencies": {
     "@auth/core": "^0.28.0",
+    "zod": "^3.22.4",
     "tsup": "6.7.0"
   }
 }
